salaService

+++ cinepolisBDA/src/main/java/persistencia/SalaDAO.java
@@ -4,7 +4,9 @@
  */
 package persistencia;
 
+import dtos.ClienteDTO;
 import dtos.SalaDTO;
+import entidad.ClienteEntidad;
 import entidad.SalaEntidad;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -16,7 +18,7 @@ import java.sql.Statement;
  *
  * @author SDavidLedesma
  */
-public class SalaDAO {
+public class SalaDAO implements ISalaDAO {
 
     private IConexionBD conexionBD;
     private Connection conexionGeneral;
@@ -25,7 +27,8 @@ public class SalaDAO {
         this.conexionBD = new ConexionBD();
     }
 
-    public void actualizarSala(SalaEntidad sala) throws SQLException {
+    @Override
+    public SalaEntidad actualizarSala(SalaEntidad sala) throws SQLException {
         String sql = "UPDATE Salas SET nombre = ?, capacidadAsientos = ?, idSucursal = ?,  WHERE idSala = ?";
         try (Connection connection = conexionBD.obtenerConexion(); PreparedStatement stmt = connection.prepareStatement(sql)) {
             stmt.setString(1, sala.getNombre());
@@ -34,18 +37,103 @@ public class SalaDAO {
             //stmt.setInt(4, sala.getIdTipoDeSala());
             stmt.setInt(4, sala.getIdSala());
             stmt.executeUpdate();
+            
+            int filasAfectadas = stmt.executeUpdate();
+
+            // Imprimimos el número de filas afectadas por la actualización
+            System.out.println("Filas afectadas: " + filasAfectadas);
+
+            // Cerramos el PreparedStatement y la conexión a la base de datos
+            stmt.close();
+            connection.close();
+
+            return this.obtenerSalaPorId(sala.getIdSala());
         }
 
     }
 
-    public void eliminarSala(int idSala) throws SQLException {
-        String sql = "DELETE FROM Salas WHERE idPelicula = ?";
-        try (Connection connection = conexionBD.obtenerConexion(); PreparedStatement stmt = connection.prepareStatement(sql)) {
-            stmt.setInt(1, idSala);
-            stmt.executeUpdate();
+    @Override
+    public SalaEntidad eliminarSala(int idSala) throws PersistenciaException {
+        
+        try {
+
+            SalaEntidad salaEliminada = this.obtenerSalaPorId(idSala);
+            if (salaEliminada == null) {
+                throw new PersistenciaException("Ocurrio un error en obtener la información del cliente por su clave");
+            }
+
+            Connection conexion = this.conexionBD.obtenerConexion();
+
+            // Definimos la consulta SQL para actualizar un cliente en la tabla 'clientes'
+            String sql = "DELETE FROM Salas WHERE idPelicula = ?";
+
+            // Creamos un objeto PreparedStatement para ejecutar la consulta de actualización
+            PreparedStatement preparedStatement = conexion.prepareStatement(sql);
+
+            // Asignamos valores a los parámetros de la consulta SQL            
+            preparedStatement.setInt(1, idSala);
+
+            // Ejecutamos la consulta de actualización en la base de datos
+            int filasAfectadas = preparedStatement.executeUpdate();
+
+            // Imprimimos el número de filas afectadas por la actualización
+            System.out.println("Filas afectadas: " + filasAfectadas);
+
+            // Cerramos el PreparedStatement y la conexión a la base de datos
+            preparedStatement.close();
+            conexion.close();
+
+            return salaEliminada;
+
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+            throw new PersistenciaException("Ocurrió un error al modificar, inténtelo de nuevo y si el error persiste comuníquese con el encargado del sistema.");
         }
     }
 
+    @Override
+    public SalaEntidad guardar(SalaDTO sala) throws PersistenciaException {
+        try {
+            Connection conexion = this.conexionBD.obtenerConexion();
+            String insert = """
+                                    INSERT INTO salas (nombre,
+                                                       capacidadAsientos,
+                                                       tiempoDeLimpiezaMinutos,
+                                                       precioActual,
+                                                       idSucursal)
+                                                 VALUES (?,?,?,?,?);
+                                    """;
+
+            PreparedStatement preparedStatement = conexion.prepareStatement(insert, Statement.RETURN_GENERATED_KEYS);
+            preparedStatement.setString(1, sala.getNombre());
+            preparedStatement.setInt(2, sala.getAsientosDisponibles());
+            preparedStatement.setInt(3, sala.getTiempoLimpiezaEnMinutos());
+            preparedStatement.setFloat(4,(float)sala.getPrecioActual());
+            preparedStatement.setInt(5, sala.getIdSala());
+
+            int filasAfectadas = preparedStatement.executeUpdate();
+            if (filasAfectadas == 0) {
+                throw new PersistenciaException("La inserción del cliente falló, no se pudo insertar el registro.");
+            }
+
+            int id = 0;
+            ResultSet resultado = preparedStatement.getGeneratedKeys();
+            if (resultado.next()) {
+                id = (resultado.getInt(1));
+            }
+
+            resultado.close();
+            preparedStatement.close();
+            conexion.close();
+
+            return this.obtenerSalaPorId(id);
+
+        } catch (SQLException ex) {
+            System.out.println(ex.getMessage());
+            throw new PersistenciaException("Ocurrió un error al leer la base de datos, inténtelo de nuevo y si el error persiste comuníquese con el encargado del sistema.");
+        }
+    }
+    
     private int guardarSala(SalaDTO sala) throws SQLException {
         int idSala = 0;
         String insertCliente = """
@@ -74,6 +162,7 @@ public class SalaDAO {
         return idSala;
     }
     
+    @Override
     public SalaEntidad guardarConTransacion(SalaDTO sala) throws PersistenciaException {
         try {
             this.conexionGeneral = this.conexionBD.obtenerConexion();
@@ -103,6 +192,7 @@ public class SalaDAO {
         }
     }
 
+    @Override
     public SalaEntidad obtenerSalaPorId(int idSala) throws SQLException {
         SalaEntidad sala = null;
         String sql = """
